+++ session_manager/views/__init__.py
@@ -1,0 +1,1 @@
+"""Views for the session manager app."""
